+++ .github/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Add `cutensornet` backed `MPS` C++ layer to `lightning.tensor`.
   [(#704)](https://github.com/PennyLaneAI/pennylane-lightning/pull/704)
 
+* Add support for `C(BlockEncode)` in Lightning devices.
+  [(#)]()
+
 ### Breaking changes
 
 * Removed the `QuimbMPS` class and the corresponding interface/backend from `lightning.tensor`.